SBT转Maven, 修复4个离线分析+实时流(Kafka累计模式), 添加项目启动指南

+++ src/main/scala/com/spark/project/analysis/member1/Analysis2.scala
@@ -30,7 +30,7 @@ object Analysis2 {
       .orderBy($"game_count".desc)
 
     val totalGames = aggregated.agg(sum("game_count")).first().getLong(0)
-    val totalOwnersAll = aggregated.agg(sum("total_owners")).first().getLong(0)
+    val totalOwnersAll = aggregated.agg(sum("total_owners")).first().getDouble(0).toLong
 
     val withPct = aggregated
       .withColumn("game_pct", round($"game_count" / totalGames * 100, 2))
